+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.6
+- add Future<IppPack> request.timeout param. return never null !!;
+
+## 1.0.5
+- add putOperationAttributes and putJobAttributes function and improve code logic
+
 ## 1.0.4
 
 - open to public _body Byte Arrays, can directly set;